remove heartbeat abstract class
