+++ src/main/java/com/actiontech/dble/manager/response/ShowHeartbeatDetail.java
@@ -8,7 +8,7 @@ package com.actiontech.dble.manager.response;
 import com.actiontech.dble.DbleServer;
 import com.actiontech.dble.backend.datasource.PhysicalDBPool;
 import com.actiontech.dble.backend.datasource.PhysicalDatasource;
-import com.actiontech.dble.backend.heartbeat.DBHeartbeat;
+import com.actiontech.dble.backend.heartbeat.MySQLHeartbeat;
 import com.actiontech.dble.backend.mysql.PacketUtil;
 import com.actiontech.dble.config.ErrorCode;
 import com.actiontech.dble.config.Fields;
@@ -110,7 +110,7 @@ public final class ShowHeartbeatDetail {
         ServerConfig conf = DbleServer.getInstance().getConfig();
         String ip = "";
         int port = 0;
-        DBHeartbeat hb = null;
+        MySQLHeartbeat hb = null;
 
         Map<String, PhysicalDBPool> dataHosts = conf.getDataHosts();
         for (PhysicalDBPool pool : dataHosts.values()) {
